analytics-service/requirements.txt
Flask
pymysql
pymongo
